tests: make the test suite link with static plugins (WebAssembly)
On WebAssembly every score plugin is a static archive, so definitions that a
shared library would have hidden behind its own copy become hard duplicate
symbols at link:

- CurveSampleTest / CurveRoundtripTest redefined the QDataStream operators for
  Curve::EasingData and Curve::PowerSegmentData that score_plugin_curve already
  exports.
- ExpressionParsingTests #included State/ExpressionParser.cpp to reach the
  Spirit grammars in its anonymous namespace, and got State::parseExpression
  with them. The grammars move to State/ExpressionParser.hpp, which both the
  translation unit and the test include; only parseExpression stays in the .cpp.

test_unit_mapping_values was declared unconditionally although the other
plugin-dependent tests are guarded; score_plugin_mapping is not part of the
WebAssembly plugin set, so the target was configured with no include paths and
failed to compile.

Finally, --bind moves from score_add_test to setup_score_common_features so
that every executable gets it, not just the ones declared through
score_add_test: Avnd_ossia_value_Test is not, and every _emval_* symbol Qt6Core
and Qt6Gui reference was undefined at link.
